move conn to request method
This moves the connection session into the request method in order to
facilitate better support for multiprocessing. In particular, a Requests
Session object used across processes will fail to properly encrypt data
via SSL[0][1]. By encapsulating the session creation in the request
method itself we workaround this issue at the limited expense of
creating the session on demand.

A version bump, from 1.1.7 to 1.1.8, is included as well.

Fixes errors that look like: [Errno 1] _ssl.c:1429

[0] http://stackoverflow.com/a/3724938/1953283
[1] https://github.com/kennethreitz/requests/issues/1906